pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-agent"
version = "0.1.0"
description = "An AI-powered coding agent built with Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "python-dotenv",
    "pydantic",
    "pyautogen",
    "volcengine-python-sdk",
    "rich",
    "autogen>=0.9.7",
    "openai>=2.14.0",
    "arize-phoenix-otel>=0.14.0",
    "arize-phoenix>=12.15.1",
    "openinference-instrumentation-openai>=0.1.41",
    "openinference-instrumentation-autogen>=0.1.10",
    "llama-index>=0.10.68",
    "llama-index-llms-openai>=0.1.27",
    "llama-index-embeddings-openai>=0.1.11",
    "llama-index-readers-file>=0.1.33",
    "llama-index-vector-stores-chroma>=0.1.10",
    "tree-sitter>=0.23.2",
    "tree-sitter-python>=0.23.6",
    "prompt-toolkit>=3.0.52",
    "pyfiglet>=1.0.4",
    "llama-index-core",
    "tree-sitter-languages>=1.10.2",
    "llama-index-llms-openai-like>=0.1.3",
    "pygls>=2.0.0",
    "python-lsp-server>=1.14.0",
    "pyright>=1.1.407",
    "watchdog>=6.0.0",
    "ripgrepy>=2.2.0",
    "python-ripgrep>=0.0.8",
    "ripgrep>=15.0.0",
    "mcp>=1.25.0",
]

[tool.setuptools.packages.find]
where = ["src"]
